requirements-web.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --extra=web --index-url=None --output-file=requirements-web.txt --pip-args=None pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
blinker==1.9.0
    # via flask
click==8.3.0
    # via flask
flask==3.1.2
    # via
    #   flask-wtf
    #   whatsthedamage (pyproject.toml)
flask-wtf==1.2.2
    # via whatsthedamage (pyproject.toml)
gunicorn==23.0.0
    # via whatsthedamage (pyproject.toml)
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
joblib==1.5.2
    # via scikit-learn
lxml==6.0.1
    # via whatsthedamage (pyproject.toml)
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
    #   wtforms
numpy==2.3.3
    # via
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via gunicorn
pandas==2.3.2
    # via whatsthedamage (pyproject.toml)
pydantic==2.11.9
    # via whatsthedamage (pyproject.toml)
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-magic==0.4.27
    # via whatsthedamage (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via whatsthedamage (pyproject.toml)
scikit-learn==1.7.2
    # via whatsthedamage (pyproject.toml)
scipy==1.16.2
    # via scikit-learn
six==1.17.0
    # via python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
werkzeug==3.1.3
    # via
    #   flask
    #   whatsthedamage (pyproject.toml)
wtforms==3.2.1
    # via
    #   flask-wtf
    #   whatsthedamage (pyproject.toml)
